refactor: refactoring projectcard ♻️
