+++ SharePointFramework/PortfolioWebParts/src/components/ProjectList/ProjectCard/ProjectCardContent/index.ts
@@ -1,0 +1,1 @@
+export * from './ProjectCardContent'
