+++ SharePointFramework/PortfolioWebParts/src/components/ProjectList/ProjectCard/ProjectCardContent/useProjectCardContent.ts
@@ -32,5 +32,5 @@ export function useProjectCardContent() {
           type: 'Prosjekttype'
         }
       })
-  } as const
+  }
 }
