+++ SharePointFramework/PortfolioWebParts/src/components/ProjectList/ProjectCard/ProjectCardFooter/index.ts
@@ -1,0 +1,1 @@
+export * from './ProjectCardFooter'
