+++ SharePointFramework/PortfolioWebParts/src/components/ProjectList/ProjectCard/ProjectCardHeader/index.ts
@@ -1,0 +1,1 @@
+export * from './ProjectCardHeader'
