+++ SharePointFramework/PortfolioWebParts/src/components/ProjectList/ProjectCard/index.ts
@@ -1,0 +1,2 @@
+export * from './ProjectCard'
+export * from './context'
